Revise LinkToGithub logic; clean up logic, revise tests
