+++ tests/middleware/test_callback_primary_logic.py
@@ -7,25 +7,19 @@ from flask import Response
 
 from database_client.enums import ExternalAccountTypeEnum
 from middleware.primary_resource_logic.callback_primary_logic import (
-    get_flask_session_callback_info,
-    get_oauth_callback_info,
-    callback_outer_wrapper,
-    callback_inner_wrapper,
-    link_github_account_request,
-    link_github_account,
-    get_github_user_info,
+    get_flask_session_callback_info
 )
+from middleware.primary_resource_logic.github_oauth_logic import link_github_account_request, link_github_account, \
+    get_github_user_info
 from middleware.custom_dataclasses import (
     FlaskSessionCallbackInfo,
     OAuthCallbackInfo,
     GithubUserInfo,
 )
-from middleware.enums import CallbackFunctionsEnum
-from middleware.primary_resource_logic.user_queries import UserRequestDTO
 from tests.helper_scripts.DynamicMagicMock import DynamicMagicMock
 
 PATCH_PREFIX = "middleware.primary_resource_logic.callback_primary_logic"
-
+GITHUB_OAUTH_PREFIX = "middleware.primary_resource_logic.github_oauth_logic"
 
 class GetFlaskSessionCallbackInfoMocks(DynamicMagicMock):
     get_callback_params: MagicMock
@@ -54,72 +48,6 @@ def test_get_flask_session_callback_info():
 class GetOauthCallbackInfoMocks(DynamicMagicMock):
     get_github_user_info: MagicMock
     get_github_oauth_access_token: MagicMock
-
-
-def test_get_oauth_callback_info():
-    mock = GetOauthCallbackInfoMocks(
-        patch_root=PATCH_PREFIX,
-        return_values={
-            "get_github_user_info": MagicMock(),
-            "get_github_oauth_access_token": MagicMock(),
-        },
-    )
-
-    result = get_oauth_callback_info()
-    mock.get_github_oauth_access_token.assert_called_once()
-    mock.get_github_user_info.assert_called_once_with(
-        mock.get_github_oauth_access_token.return_value
-    )
-
-    assert isinstance(result, OAuthCallbackInfo)
-
-    assert result.github_user_info == mock.get_github_user_info.return_value
-
-
-class CallbackOuterWrapperMocks(DynamicMagicMock):
-    callback_inner_wrapper: MagicMock
-    get_oauth_callback_info: MagicMock
-    get_flask_session_callback_info: MagicMock
-
-
-def test_callback_outer_wrapper():
-
-    mock = CallbackOuterWrapperMocks(
-        patch_root=PATCH_PREFIX,
-        return_values={
-            "callback_inner_wrapper": MagicMock(spec=Response),
-            "get_oauth_callback_info": OAuthCallbackInfo(
-                github_user_info=MagicMock(spec=GithubUserInfo)
-            ),
-            "get_flask_session_callback_info": FlaskSessionCallbackInfo(
-                callback_functions_enum=MagicMock(spec=CallbackFunctionsEnum),
-                callback_params=MagicMock(spec=dict),
-            ),
-        },
-    )
-
-    result = callback_outer_wrapper(mock.db_client)
-
-    mock.get_oauth_callback_info.assert_called_once()
-    mock.get_flask_session_callback_info.assert_called_once()
-    mock.callback_inner_wrapper.assert_called_once_with(
-        db_client=mock.db_client,
-        callback_function_enum=mock.get_flask_session_callback_info.return_value.callback_functions_enum,
-        github_user_info=mock.get_oauth_callback_info.return_value.github_user_info,
-        callback_params=mock.get_flask_session_callback_info.return_value.callback_params,
-    )
-
-    assert isinstance(result, Response)
-
-
-class CreateUserWithGithubMocks(DynamicMagicMock):
-    user_post_results: MagicMock
-    create_random_password: MagicMock
-    link_github_account: MagicMock
-    make_response: MagicMock
-
-
-
 
 
 class CallbackInnerWrapperMocks(DynamicMagicMock):
@@ -155,69 +83,6 @@ def setup_callback_inner_wrapper_mocks():
     return mock
 
 
-def run_and_validate_callback_inner_wrapper_with_mocks(
-    mock: CallbackInnerWrapperMocks,
-    callback_function_enum: CallbackFunctionsEnum,
-    check_is_response: bool = True,
-):
-    result = callback_inner_wrapper(
-        db_client=mock.db_client,
-        callback_function_enum=callback_function_enum,
-        github_user_info=mock.github_user_info,
-        callback_params=mock.callback_params,
-    )
-    if check_is_response:
-        assert isinstance(result, Response)
-
-
-def test_callback_inner_wrapper_login_with_github(setup_callback_inner_wrapper_mocks):
-
-    mock = setup_callback_inner_wrapper_mocks
-
-    run_and_validate_callback_inner_wrapper_with_mocks(
-        mock=mock,
-        callback_function_enum=CallbackFunctionsEnum.LOGIN_WITH_GITHUB,
-    )
-
-    assert_callback_inner_wrapper_function_calls(
-        mock=mock,
-        called_function="try_logging_in_with_github_id",
-        db_client=mock.db_client,
-        github_user_info=mock.github_user_info,
-    )
-
-
-def test_callback_inner_wrapper_link_to_github(setup_callback_inner_wrapper_mocks):
-
-    mock = setup_callback_inner_wrapper_mocks
-
-    run_and_validate_callback_inner_wrapper_with_mocks(
-        mock=mock,
-        callback_function_enum=CallbackFunctionsEnum.LINK_TO_GITHUB,
-    )
-
-    assert_callback_inner_wrapper_function_calls(
-        mock=mock,
-        called_function="link_github_account_request",
-        db_client=mock.db_client,
-        github_user_info=mock.github_user_info,
-        pdap_account_email=mock.callback_params["user_email"],
-    )
-
-
-def test_callback_inner_wrapper_invalid_callback_function_enum(
-    setup_callback_inner_wrapper_mocks,
-):
-    mock = setup_callback_inner_wrapper_mocks
-
-    with pytest.raises(ValueError):
-        run_and_validate_callback_inner_wrapper_with_mocks(
-            mock=mock, callback_function_enum=MagicMock(), check_is_response=False
-        )
-
-    assert_callback_inner_wrapper_function_calls(mock=mock, called_function=None)
-
-
 class LinkGithubAccountRequestMocks(DynamicMagicMock):
     link_github_account: MagicMock
     make_response: MagicMock
@@ -226,7 +91,7 @@ class LinkGithubAccountRequestMocks(DynamicMagicMock):
 def test_link_github_account_request():
 
     mock = LinkGithubAccountRequestMocks(
-        patch_root=PATCH_PREFIX,
+        patch_root=GITHUB_OAUTH_PREFIX,
     )
 
     link_github_account_request(
@@ -251,7 +116,7 @@ class LinkGithubAccountMocks(DynamicMagicMock):
 
 def test_link_github_account():
 
-    mock = LinkGithubAccountMocks(patch_root=PATCH_PREFIX)
+    mock = LinkGithubAccountMocks(patch_root=GITHUB_OAUTH_PREFIX)
     mock.github_user_info.user_email = mock.pdap_account_email
     mock.db_client.get_user_info.return_value = mock.db_client_user_info
 
@@ -278,7 +143,7 @@ class GetGithubUserInfoMocks(DynamicMagicMock):
 def test_get_github_user_info():
 
     mock = GetGithubUserInfoMocks(
-        patch_root=PATCH_PREFIX,
+        patch_root=GITHUB_OAUTH_PREFIX,
         return_values={
             "get_github_user_id": MagicMock(),
             "get_github_user_email": MagicMock(),
